+++ src/pymodule/py_pingtools/c_beamsampleselection.cpp
@@ -28,12 +28,12 @@ void init_c_beamsampleselection(pybind11::module& m)
 {
     py::class_<BeamSampleSelection, BeamSelection, std::shared_ptr<BeamSampleSelection>>(
         m,
-        "BeamSampleSelection",
+        "BeamSampleSelection", 
         DOC(themachinethatgoesping, echosounders, pingtools, BeamSampleSelection))
-        .def(py::init<uint16_t>(),
+        .def(py::init<uint32_t>(),
              DOC_BeamSampleSelection(BeamSampleSelection),
              py::arg("sample_step_ensemble") = 1)
-        .def(py::init<std::vector<uint16_t>, std::vector<uint16_t>, uint16_t>(),
+        .def(py::init<std::vector<uint32_t>, std::vector<uint32_t>, uint32_t>(),
              DOC_BeamSampleSelection(BeamSampleSelection_2),
              py::arg("first_sample_number_per_beam"),
              py::arg("last_sample_number_per_beam"),
